Getting Accession Data from ICARDA's Genebank Documentation System

Return a data frame with accession data for the specified crop.

Usage

getAccessions(
  crop = "",
  ori = NULL,
  IG = "",
  doi = FALSE,
  taxon = FALSE,
  collectionYear = FALSE,
  coor = FALSE,
  available = FALSE
)

Arguments

crop

character. Crop for which to get accession data. See section 'Details' for available crops or use getCrops function. Default: "".

ori

string. Country of origin using the ISO 3166-1 alpha-3 country codes. Default: NULL.

IG

integer. Unique identifier of accession. Default: "".

doi

boolean. If TRUE , the function will return the digital object identifiers DOI for the accessions. Default: FALSE.

taxon

boolean. If TRUE, the function will return the taxon information of the accessions. Default: FALSE.

collectionYear

boolean. If TRUE, the function will return the year of the collecting mission. Default: FALSE.

coor

boolean. If TRUE, returns only georeferenced accessions containing longitude and latitude. Default: FALSE.

available

boolean. If TRUE, returns the availability of accessions for distribution, Default: FALSE.

Value

A data frame with accession passport data for specified crop in crop from the locations in ori.

Details

Types of crops available include:

'ICAG'

Aegilops

'ICB'

Barley

'ICBW'

Bread wheat

'ILC'

Chickpea

'ICDW'

Durum wheat

'ILB'

Faba bean

'BPL'

Faba bean BPL

'IFMI'

Forage and range

'IFLA'

Lathyrus

'ILL'

Lentil

'IFMA'

Medicago annual

'IC'

Not mandate cereals

'IFPI'

Pisum

'ICPW'

Primitive wheat

'IFTR'

Trifolium

'IFVI'

Vicia

'ICWH'

Wheat hybrids

'ICWW'

Wheat wild relatives

'ILWC'

Wild Cicer

'ICWB'

Wild Hordeum

'ILWL'

Wild Lens

'ICWT'

Wild Triticum

Alternatively, the list of available crops can be fetched from ICARDA's online server using getCrops.

Author

Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar

Examples

if (FALSE) { # \dontrun{
 # Obtain accession data for durum wheat
 durum <- getAccessions(crop = 'Durum wheat', coor = TRUE)
 } # }