Getting Trait Values of Accessions for a Specific Trait

Return a data frame with observed values of accessions for associated Trait

Usage

getTraitsData(IG, traitID)

Arguments

IG

factor. Unique identifier of accession.

traitID

integer. Unique identifier of trait (from getTraits).

Value

A data frame with scores for the trait specified in traitID for the accessions given in IG.

Details

Possible inputs for traitID can be found using the getTraits function (see section 'Examples').

Author

Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 # Check trait ID for septoria and get septoria data for durum wheat
 durum <- getAccessions(crop = 'Durum wheat', coor = TRUE)
 durumTraits <- getTraits(crop = 'Durum wheat')
 septoria <- getTraitsData(IG = durum$IG, traitID = 145)
 }
} # }