Extracting historical climate data from WorldClim 2.1

extractWCdata returns a data frame based on specified climatic variables. This function modifies the getData function from the raster R package to extract world climate data of version 2.1 instead of version 1.4.

Usage

extractWCdata(sites, long, lat, var, res = 2.5)

Arguments

sites

object of class "data.frame" with coordinates of sites from which to extract data.

long

character. Name of column from sites with longitude.

lat

character. Name of column from sites with latitude.

var

character. Climatic variable(s) to be extracted: 'tavg', 'tmin', 'tmax', 'prec', 'bio', 'srad', 'vapr', 'wind'.

res

numeric. Spatial resolution. Default 2.5.

Value

An object of class "data.frame" with specified climatic variables for coordinates in sites.

Details

A grid can be created with any particular coordinates and used as input for sites (see section 'Examples'). extractWCdata will use the given coordinates to extract data from the WorldClim 2.1 database. The extracted data will most likely contain NAs for sites where climate data is not available. These should be removed or imputed before using the data to make predictions.

Author

Zakaria Kehel, Fawzy Nawar, Bancy Ngatia, Khadija Aouzal, Chafik Analy

Examples

if (FALSE) { # \dontrun{
 # Create grid
 long <- seq(-16, 115, length = 3)
 lat <- seq(25, 59, length = 3)
 sf <- expand.grid(x = sp1, y = sp2)

 # Extract data using grid
 sp.df0 <- extractWCdata(sf, long = 'long', lat = 'lat', var = 'bio')
 sp.df <- na.omit(sp.df0)
 } # }