Getting Accession Data from ICARDA's Genebank Database.

Return a data frame with accession data for the specified crop or specified list of accession numbers (i.e. IGs).

Usage

getAccessions(
  crop = "",
  ori = NULL,
  IG = "",
  doi = FALSE,
  taxon = FALSE,
  collectionYear = FALSE,
  coor = FALSE,
  available = FALSE,
  other_id = FALSE
)

Arguments

crop

character. Crop for which to get accession data. See section 'Details' for available crops or use getCrops function. Default: "".

ori

string. Country of origin using the ISO 3166-1 alpha-3 country codes. Default: NULL.

IG

integer. List of accession numbers. Default: "".

doi

boolean. If TRUE , the function will return the digital object identifiers DOI for the accessions. Default: FALSE.

taxon

boolean. If TRUE, the function will return the taxon information of the accessions. Default: FALSE.

collectionYear

boolean. If TRUE, the function will return the year of the collecting mission. Default: FALSE.

coor

boolean. If TRUE, returns only georeferenced accessions. Default: FALSE.

available

boolean. If TRUE, returns only available accessions for distribution, Default: FALSE.

Value

A data frame with accession passport data for specified crop in crop from the locations in ori if specified.

Details

The list of available crops can be fetched from ICARDA's Genebank database using getCrops.

Author

Khadija Aouzal, Amal Ibnelhobyb, Zakaria Kehel, Fawzy Nawar

Examples

if (FALSE) { # \dontrun{
 # Obtain accession data for durum wheat
 durum <- getAccessions(crop = 'Durum wheat', coor = TRUE)
 } # }