Splitting the Data, Tuning and Training the models, and making predictions

Automatic function for tuning and training data, it returns a list containing a model objects, data frame and plot.

Usage

tuneTrain(
  data,
  y,
  p = 0.7,
  method = "knn",
  imbalanceMethod = NULL,
  imbalanceThreshold = 0.2,
  parallelComputing = FALSE,
  control = "repeatedcv",
  Length = 10,
  number = 10,
  repeats = 10,
  process = c("center", "scale"),
  summary = multiClassSummary,
  classProbs = FALSE,
  ...
)

Arguments

data

object of class "data.frame" with target variable and predictor variables.

y

character. Target variable.

p

numeric. Proportion of data to be used for training. Default: 0.7

method

character. Type of model to use for classification or regression.

imbalanceMethod

Method for handling imbalanced data ("up", "down"). Default: NULL.

imbalanceThreshold

Threshold to determine if data is imbalanced (numeric between 0 and 0.4). Default: 0.2.

parallelComputing

logical. indicates whether to also use the parallel processing. Default: False

control

character. Resampling method to use. Choices include: "boot", "boot632", "optimism_boot", "boot_all", "cv", "repeatedcv", "LOOCV", "LGOCV", "none", "oob", timeslice, "adaptive_cv", "adaptive_boot", or "adaptive_LGOCV". Default: "repeatedcv". See train for specific details on the resampling methods.

Length

integer. Number of values to output for each tuning parameter. If search = "random" is passed to trainControl through ..., this becomes the maximum number of tuning parameter combinations that are generated by the random search. Default: 10.

number

integer. Number of cross-validation folds or number of resampling iterations. Default: 10.

repeats

integer. Number of folds for repeated k-fold cross-validation if "repeatedcv" is chosen as the resampling method in control. Default: 10.

process

character. Defines the pre-processing transformation of predictor variables to be done. Options are: "BoxCox", "YeoJohnson", "expoTrans", "center", "scale", "range", "knnImpute", "bagImpute", "medianImpute", "pca", "ica", or "spatialSign". See preProcess for specific details on each pre-processing transformation. Default: c('center', 'scale').

summary

expression. Computes performance metrics across resamples. For numeric y, the mean squared error and R-squared are calculated. For factor y, the overall accuracy and Kappa are calculated. See trainControl and defaultSummary for details on specification and summary options. Default: multiClassSummary.

classProbs

a logical (default : TRUE assuming classification task). It should be set to FALSE for regression tasks in conjunction with summary argument set to "defaultSummary". See examples for more details.

...

additional arguments to be passed to createDataPartition, trainControl and train functions in the package caret.

Value

A list object with results from tuning and training the model selected in method, together with predictions and class probabilities. The training and test data sets obtained from splitting the data are also returned in addition to model quality metrics and preformance plots.

If y is factor, class probabilities are calculated for each class. If y is numeric, predicted values are calculated.

A ROC curve is created if y is factor. Otherwise, a plot of residuals versus predicted values is created if y is numeric.

tuneTrain relies on packages caret, ggplot2 and pROC to perform the modelling and plotting.

Details

Types of classification and regression models available for use with tuneTrain can be found using names(getModelInfo()). The results given depend on the type of model used.

In addition to Model object, Model quality, Tuning and training datasets,For classification models, class probabilities and ROC curve are given in the results. For regression models, Variable importance, predictions and residuals versus predicted plot are given. y should be converted to either factor if performing classification or numeric if performing regression before specifying it in tuneTrain.

Imbalance handling methods "up" and "down" use ,respectively, upsample() and downsample() from caret subsampling methods.

See also

Author

Zakaria Kehel, Bancy Ngatia, Khadija Aziz, Chafik Analy

Examples